fix(typing): make the gated annotations correct across the supported numpy range
The mypy step added in the previous commit passed locally and failed on every py3.10 leg
of CI. The cause was not the interpreter: it was numpy. Local numpy is 2.5.1, CI's py3.10
resolves 2.2.6, and the two ship different stub shapes. Same mypy (2.3.0) both sides.

Ten errors, all real annotation weaknesses rather than stub noise:

- Eight `Missing type arguments for generic type "ndarray"` in distort.py. A bare
  `np.ndarray` is an implicit Any under strict checking; numpy 2.5's stubs supply defaults
  that make it pass, 2.2's do not. Since pyproject declares `numpy>=1.23`, the annotation
  that only type-checks on the newest numpy is the one that is wrong. Now spelled
  `NDArray[Any]` -- explicit, no narrowing, no runtime effect, and correct on both.
- `simulate.py:492`/`:498`: `np.linspace` is inferred as `floating[Any]` under 2.2 and
  `float64` under 2.5, so passing its result to `_lorentzian_sum` was an arg-type error on
  the older stubs. `dtype=np.float64` is now explicit; linspace already returned float64 at
  runtime, so nothing changes except that the annotation and the value now agree.

Verified in a probe venv built to match CI exactly (python 3.10, numpy 2.2.6, mypy 2.3.0):
`Success: no issues found in 2 source files`, and still green on the local 3.12/2.5.1 pair.
Checking only one numpy version is what let this through the first time.

389 passed / 12 skipped / 33 deselected; ruff clean.

+++ moldetr/simulate.py
@@ -487,7 +487,7 @@ def simulate(
             )
         amps = amps * (n_spins / total)
 
-    ppm_axis = np.linspace(float(left_ppm), float(right_ppm), n_points)
+    ppm_axis = np.linspace(float(left_ppm), float(right_ppm), n_points, dtype=np.float64)
     hz_axis = ppm_axis * float(base_freq_mhz)
     spectrum = _lorentzian_sum(hz_axis, freqs, amps, gamma)
